Fix tests
Using __METHOD__ for the routes being tested is resulting in an error now:

```
Symfony\Component\HttpFoundation\Exception\BadRequestException: Invalid URI: A URI cannot contain a backslash.
```

Not sure if this is a PHP 8.4 change or a Symfony change, but fair enough. `__FUNCTION__` returns just the method name, which should be sufficient.